Refactor: Robust forecast step logic, B/T chart notation, PDF pagination, and interactive payload handling
This commit introduces major improvements across analysis, reporting, charting, and API layers to enhance accuracy, performance, and user experience for large datasets and advanced reporting scenarios.

Key changes:
- Forecast step calculation unified and now per-column, matching forecast_pct visual share and handling NaNs.
- Chart labeling and axis formatting use B/T notation for large values in all chart types.
- PDF report generation overhauled for pagination, layout, and robust AI summary/table rendering.
- Interactive API and payloads downsampled for large datasets, with async fallback and anomaly marker improvements.
- Model attribution and AI summary fallback logic improved, with legacy marker backfill.
- Static plots ZIP export uses forecast_pct and per-column history, with improved chart labeling and logging.
- Template/frontend enhancements for chart rendering and payload handling.
- Comprehensive new and updated tests for all major changes.

Motivation: These changes address performance bottlenecks, improve clarity for large numeric datasets, ensure consistent forecast visualization, and provide robust reporting and charting for advanced analysis workflows. All modifications are covered by new and updated tests to ensure reliability and prevent regressions.

+++ data_analysis/core/cache.py
@@ -13,7 +13,7 @@ _MISSING = object()
 class TinyLRU(OrderedDict[Any, Any]):
     """Small LRU cache with optional memory-size based eviction."""
 
-    def __init__(self, max_items: int = 6, max_size_mb: int | None = None):
+    def __init__(self, max_items: int = 6, max_size_mb: float | None = None):
         super().__init__()
         self.max_items = max_items
         self.max_size_mb = max_size_mb
